revise discoverability of tooltip
to cut down on over-announcements with chrome, and to cut down on the possibility of people hearing the announcement with the owning element, and then again in isolation after navigating away from the element with the VC, set the tooltip to be aria-hidden=true.

this commit may be undone depending on the direction taken to clarify the usefulness of `role=tooltip` in the ARIA specification.